updated files and containerized application
