+++ Quic-Cal-Agent/date_server/requirements.txt
@@ -1,0 +1,2 @@
+fastmcp==2.7.1
+tzdata
