+++ Quic-Cal-Agent/execute.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+
+
+
+
+# kill -9 $(lsof -t -i:8123)
+# kill -9 $(lsof -t -i:4200)
+# kill -9 $(lsof -t -i:4201)
+
+# open -a Docker
+
+docker compose up --build
