+++ Quic-Cal-Agent/gcal_server/requirements.txt
@@ -1,0 +1,1 @@
+fastmcp==2.7.1
