+++ Quic-Cal-Agent/src/agent/.gitignore
@@ -2,6 +2,8 @@
 __pycache__/
 *.py[cod]
 *$py.class
+credentials.json
+.env
 
 # C extensions
 *.so
