+++ Quic-Cal-Agent/src/agent/requirements.txt
@@ -1,0 +1,127 @@
+annotated-types==0.7.0
+fastmcp==2.7.1
+anthropic==0.51.0
+anyio==4.8.0
+appnope==0.1.4
+asttokens==3.0.0
+blockbuster==1.5.24
+cachetools==5.5.1
+certifi==2024.12.14
+cffi==1.17.1
+charset-normalizer==3.4.1
+click==8.2.1
+cloudpickle==3.1.1
+comm==0.2.2
+cryptography==44.0.3
+debugpy==1.8.11
+decorator==5.1.1
+distro==1.9.0
+exceptiongroup==1.2.2
+executing==2.1.0
+forbiddenfruit==0.1.4
+fuzzywuzzy==0.18.0
+google-api-core==2.24.1
+google-api-python-client==2.171.0
+google-auth==2.38.0
+google-auth-httplib2==0.2.0
+google-auth-oauthlib==1.2.2
+googleapis-common-protos==1.66.0
+grpcio==1.70.0
+h11==0.14.0
+httpcore==1.0.7
+httplib2==0.22.0
+httpx==0.28.1
+httpx-sse==0.4.0
+idna==3.10
+importlib_metadata==8.6.1
+ipykernel==6.29.5
+ipython==8.31.0
+jedi==0.19.2
+jiter==0.8.2
+jsonpatch==1.33
+jsonpointer==3.0.0
+jsonschema_rs==0.29.1
+jupyter_client==8.6.3
+jupyter_core==5.7.2
+langchain==0.3.25
+langchain-anthropic==0.3.13
+langchain-core==0.3.64
+langchain-mcp-adapters==0.1.4
+langchain-openai==0.3.17
+langchain-text-splitters==0.3.8
+langgraph==0.4.5
+langgraph-api==0.2.42
+langgraph-checkpoint==2.0.26
+langgraph-cli==0.2.11
+langgraph-prebuilt==0.1.8
+langgraph-runtime-inmem==0.2.1
+langgraph-sdk==0.1.69
+langsmith==0.3.45
+matplotlib-inline==0.1.7
+mcp==1.9.2
+nest_asyncio==1.6.0
+oauthlib==3.2.2
+ollama==0.4.7
+openai==1.79.0
+orjson==3.10.16
+ormsgpack==1.9.1
+packaging==24.2
+parso==0.8.4
+pexpect==4.9.0
+pickleshare==0.7.5
+pinecone==6.0.1
+pinecone-plugin-interface==0.0.7
+pip==24.2
+platformdirs==4.3.6
+prompt_toolkit==3.0.50
+proto-plus==1.26.0
+protobuf==5.29.3
+protoc-gen-openapiv2==0.0.1
+psutil==6.1.1
+ptyprocess==0.7.0
+pure_eval==0.2.3
+pyasn1==0.6.1
+pyasn1_modules==0.4.1
+pycparser==2.22
+pydantic==2.10.5
+pydantic_core==2.27.2
+pydantic-settings==2.9.1
+Pygments==2.19.1
+PyJWT==2.10.1
+pyparsing==3.2.1
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.0
+python-multipart==0.0.20
+PyYAML==6.0.2
+pyzmq==26.2.0
+regex==2024.11.6
+requests==2.32.3
+requests-oauthlib==2.0.0
+requests-toolbelt==1.0.0
+rsa==4.9
+setuptools==75.1.0
+six==1.17.0
+sniffio==1.3.1
+SQLAlchemy==2.0.41
+sse-starlette==2.1.3
+stack_data==0.6.3
+starlette==0.47.0
+structlog==25.4.0
+tenacity==9.1.2
+tiktoken==0.9.0
+tornado==6.4.2
+tqdm==4.67.1
+traitlets==5.14.3
+truststore==0.10.1
+typing_extensions==4.12.2
+typing-inspection==0.4.1
+uritemplate==4.1.1
+urllib3==2.3.0
+uvicorn==0.34.2
+watchfiles==1.0.5
+wcwidth==0.2.13
+websockets==14.2
+wheel==0.44.0
+xxhash==3.5.0
+zipp==3.21.0
+zstandard==0.23.0
