--- langgraph_agent_deployment/app/.env.example
@@ -1,4 +1,0 @@
-# To separate your traces from other application
-LANGSMITH_PROJECT=new-agent
-
-# Add API keys for connecting to LLM providers, data sources, and other integrations here
