--- langgraph_agent_deployment/app/src/agent/output.json
@@ -1,166 +1,0 @@
-{
-    "messages": [
-        {
-            "lc": 1,
-            "type": "constructor",
-            "id": [
-                "langchain",
-                "schema",
-                "messages",
-                "HumanMessage"
-            ],
-            "kwargs": {
-                "content": "what is today's date?",
-                "type": "human",
-                "id": "1b564965-9a88-43a9-98b0-1f6db029caef"
-            }
-        },
-        {
-            "lc": 1,
-            "type": "constructor",
-            "id": [
-                "langchain",
-                "schema",
-                "messages",
-                "AIMessage"
-            ],
-            "kwargs": {
-                "content": "",
-                "additional_kwargs": {
-                    "tool_calls": [
-                        {
-                            "id": "call_4tqembb5d2fjfNRhd2G3xhBl",
-                            "function": {
-                                "arguments": "{}",
-                                "name": "get_datetime"
-                            },
-                            "type": "function"
-                        }
-                    ],
-                    "refusal": null
-                },
-                "response_metadata": {
-                    "token_usage": {
-                        "completion_tokens": 10,
-                        "prompt_tokens": 69,
-                        "total_tokens": 79,
-                        "completion_tokens_details": {
-                            "accepted_prediction_tokens": 0,
-                            "audio_tokens": 0,
-                            "reasoning_tokens": 0,
-                            "rejected_prediction_tokens": 0
-                        },
-                        "prompt_tokens_details": {
-                            "audio_tokens": 0,
-                            "cached_tokens": 0
-                        }
-                    },
-                    "model_name": "gpt-4.1-2025-04-14",
-                    "system_fingerprint": "fp_b3f1157249",
-                    "id": "chatcmpl-BemcLfH0up6qoLqWGmmnfhqizn6Wg",
-                    "service_tier": "default",
-                    "finish_reason": "tool_calls",
-                    "logprobs": null
-                },
-                "type": "ai",
-                "id": "run--338a9938-7970-46df-8c8c-ff039e6b111d-0",
-                "tool_calls": [
-                    {
-                        "name": "get_datetime",
-                        "args": {},
-                        "id": "call_4tqembb5d2fjfNRhd2G3xhBl",
-                        "type": "tool_call"
-                    }
-                ],
-                "usage_metadata": {
-                    "input_tokens": 69,
-                    "output_tokens": 10,
-                    "total_tokens": 79,
-                    "input_token_details": {
-                        "audio": 0,
-                        "cache_read": 0
-                    },
-                    "output_token_details": {
-                        "audio": 0,
-                        "reasoning": 0
-                    }
-                },
-                "invalid_tool_calls": []
-            }
-        },
-        {
-            "lc": 1,
-            "type": "constructor",
-            "id": [
-                "langchain",
-                "schema",
-                "messages",
-                "ToolMessage"
-            ],
-            "kwargs": {
-                "content": "\"2025-06-04T13:55:06.173360-04:00\"",
-                "type": "tool",
-                "name": "get_datetime",
-                "id": "f6949c44-e82f-49eb-b6ec-56509e8f39fa",
-                "tool_call_id": "call_4tqembb5d2fjfNRhd2G3xhBl",
-                "status": "success"
-            }
-        },
-        {
-            "lc": 1,
-            "type": "constructor",
-            "id": [
-                "langchain",
-                "schema",
-                "messages",
-                "AIMessage"
-            ],
-            "kwargs": {
-                "content": "Today's date is June 4, 2025.",
-                "additional_kwargs": {
-                    "refusal": null
-                },
-                "response_metadata": {
-                    "token_usage": {
-                        "completion_tokens": 12,
-                        "prompt_tokens": 108,
-                        "total_tokens": 120,
-                        "completion_tokens_details": {
-                            "accepted_prediction_tokens": 0,
-                            "audio_tokens": 0,
-                            "reasoning_tokens": 0,
-                            "rejected_prediction_tokens": 0
-                        },
-                        "prompt_tokens_details": {
-                            "audio_tokens": 0,
-                            "cached_tokens": 0
-                        }
-                    },
-                    "model_name": "gpt-4.1-2025-04-14",
-                    "system_fingerprint": "fp_b3f1157249",
-                    "id": "chatcmpl-BemcNoxj6tqaeK4jDCAb3OAgwKaC1",
-                    "service_tier": "default",
-                    "finish_reason": "stop",
-                    "logprobs": null
-                },
-                "type": "ai",
-                "id": "run--9047fa38-9072-46f1-9fbd-3dea42d75da7-0",
-                "usage_metadata": {
-                    "input_tokens": 108,
-                    "output_tokens": 12,
-                    "total_tokens": 120,
-                    "input_token_details": {
-                        "audio": 0,
-                        "cache_read": 0
-                    },
-                    "output_token_details": {
-                        "audio": 0,
-                        "reasoning": 0
-                    }
-                },
-                "tool_calls": [],
-                "invalid_tool_calls": []
-            }
-        }
-    ]
-}
