--- langgraph_agent_deployment/app/tests/conftest.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.fixture(scope="session")
-def anyio_backend():
-    return "asyncio"
