--- langgraph_agent_deployment/app/tests/integration_tests/test_graph.py
@@ -1,12 +1,0 @@
-import pytest
-
-from agent import graph
-
-pytestmark = pytest.mark.anyio
-
-
-@pytest.mark.langsmith
-async def test_agent_simple_passthrough() -> None:
-    inputs = {"changeme": "some_val"}
-    res = await graph.ainvoke(inputs)
-    assert res is not None
