--- langgraph_agent_deployment/app/tests/unit_tests/test_configuration.py
@@ -1,9 +1,0 @@
-from langgraph.pregel import Pregel
-
-from agent.graph import graph
-
-
-def test_placeholder() -> None:
-    # TODO: You can add actual unit tests
-    # for your graph and other logic here.
-    assert isinstance(graph, Pregel)
